FEAT: SongRank 도메인 자료형 수정

+++ server/rhythm4cuts/src/main/java/com/b109/rhythm4cuts/model/domain/SongRank.java
@@ -13,7 +13,7 @@ public class SongRank {
 
     @Id @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "song_rank_seq")
-    private int songRankSeq;
+    private Long songRankSeq;
 
     // 노래 제목
     private String title;
